requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
black==24.10.0
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
ffmpeg==1.4
filelock==3.16.1
Flask==3.1.0
fsspec==2024.12.0
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
llvmlite==0.43.0
lxml==5.3.0
lz4==4.3.3
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai==1.59.7
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==4.25.5
protoc-gen-openapiv2==0.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
striprtf==0.0.28
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
youtube-transcript-api==0.6.3
